[REFACTOR] #12 코드 리뷰 사항 적용
